fetchinf all data done

+++ lib/models/wind.dart
@@ -1,5 +1,5 @@
 class Wind {
-  double? speed;
+  int? speed;
   int? deg;
 
   Wind({this.speed, this.deg});
